Added a separate shared directory
