+++ src/shared/App.tsx
@@ -1,4 +1,4 @@
-import MainPage from "./pages/MainPage.tsx";
+import MainPage from "../layout/MainPage.tsx";
 import {BrowserRouter} from "react-router-dom";
 
 function App() {
