+++ src/shared/components/Example.tsx
