+++ src/shared/variables/appRoutes.ts
@@ -1,4 +1,4 @@
-export const ROUTES = {
+export const APP_ROUTES = {
     HOME: '/',
     SIGN_IN: '/sign-in',
 };
