+++ src/shared/vite-env.d.ts
